Shuffle some code
- Move mem* family of functions out of x86
- Move identity mapping first few MiB into paging_init
- Remove passing bootloader PML4
