+++ kernel/src/arch/x86_64/klib/klib.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "../../../memory.h"
+
 #include <stddef.h>
 #include <stdint.h>
 
@@ -20,23 +22,12 @@ __attribute__((interrupt)) void trap(struct interrupt_frame *frame);
 
 // Initialises the heap by mapping it in virtual memory.
 void heap_init(void);
+
 // Allocates memory from a specified size.
 void *kmalloc(size_t size);
+
 // Frees the memory allocated from heap.
 void free(void *address);
-
-// Compare each byte of the two addresses given the count
-int memcmp(const void *a, const void *b, size_t count);
-
-// Sets each byte in destination given the value and size
-void memset(void *dst, char val, size_t size);
-
-// Copies the memory from source to destination given the length
-void *memcpy(void *dst, const void *src, size_t len);
-
-// Moves the memory from source to destination given the length
-// (Safe for overlapping)
-void *memmove(void *dst, const void *src, size_t len);
 
 // Exception Handler
 __attribute__((interrupt)) void exception_handler(struct interrupt_frame *frame);
