+++ kernel/src/arch/x86_64/memory/memmap.c
@@ -5,21 +5,21 @@
 
 #include <stddef.h>
 
-void memmap_print_entries(
-        uint32_t entry_count,
-        struct address_range_descriptor address_range_descriptor[]) {
-    for (size_t i = 0; i < entry_count; i++) {
+void memmap_print_entries(struct address_ranges address_ranges) {
+    for (size_t i = 0; i < address_ranges.count; i++) {
+        struct address_range_descriptor entry = address_ranges.entries[i];
         krintf(
                 "entry count %d\n"
                 "base address lower: %x, base address higher: %x\n"
                 "length lower: %x, length higher: %x\n"
                 "type: %d, extended attributes: %x\n",
-                i, address_range_descriptor[i].base_address_lower,
-                address_range_descriptor[i].base_address_higher,
-                address_range_descriptor[i].length_lower,
-                address_range_descriptor[i].length_higher,
-                address_range_descriptor[i].type,
-                address_range_descriptor[i].extended_attributes);
+                i,
+                entry.base_address_lower,
+                entry.base_address_higher,
+                entry.length_lower,
+                entry.length_higher,
+                entry.type,
+                entry.extended_attributes);
         ksleep(276447232);
     }
 }
