+++ kernel/src/arch/x86_64/tests/memory_tests.c
@@ -1,10 +1,9 @@
 #include "../../../io.h"
 #include "../../../test.h"
 #include "../memory/paging.h"
-#include <stdint.h>
 
 static void mmap_test(void) {
-    void *test_virtual_address = kmap_page(0xFFFFFFF, PAGE_PRESENT);
+    void *test_virtual_address = kmap_page((void *) 0xFFFFFFF, PAGE_PRESENT);
 
     if (!verify_mapping(test_virtual_address)) {
         puts("Could not page correctly!\n");
